server/src/main/java/br/com/allcool/news/resource/NewsResource.java
package br.com.allcool.news.resource;

import br.com.allcool.dto.NewsDTO;
import br.com.allcool.news.service.NewsService;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

import java.util.UUID;

@RestController
@RequestMapping("/api/news")
public class NewsResource {

    private final NewsService service;

    public NewsResource(NewsService service) {
        this.service = service;
    }

    @GetMapping("/{id}")
    public ResponseEntity<NewsDTO> findById(@PathVariable("id") UUID id) {

        return ResponseEntity.ok(this.service.findById(id));
    }
}
